requirements.txt
alembic==1.0.11
astroid==2.2.5
Click==7.0
colorama==0.4.1
coverage==4.5.4
Flask==1.0.3
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lark-parser==0.8.5
lazy-object-proxy==1.3.1
Mako==1.0.14
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.3.1
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==5.1.2
six==1.12.0
SQLAlchemy==1.3.5
typed-ast==1.3.5
Werkzeug==0.15.4
wrapt==1.11.1
WTForms==2.2.1
